# staff/swagger.yaml
swagger: '2.0'
info:
  version: 0.0.1
  title: Hello World App
host: 'localhost:3000'
basePath: /
schemes:
  - http
  - https
consumes:
  - application/json
produces:
  - application/json
securityDefinitions:
  APIKeyHeader:
    type: apiKey
    in: header
    name: X-API-Key
  JWT:
    description: 'Authorization: Bearer {token}'
    type: apiKey
    name: Authorization
    in: header
security:
  - APIKeyHeader: []
    JWT: []
paths:
...
  /staffs:
    post:
      summary: 创建单个员工
      description: 创建单个员工...
      parameters:
        - in: body
          name: body
          description: 模型参数
          required: true
          schema:
            $ref: '#/definitions/Staff'
      responses:
        '200':
          description: 创建成功
    get:
      summary: 查询多个员工
      description: 查询多个员工...
      parameters:
        - in: query
          name: pageNum
          type: integer
          description: 页码
          default: 1
        - in: query
          name: pageSize
          type: integer
          description: 分页大小
          default: 10
        - in: query
          name: sorter
          type: string
          description: 排序
      responses:
        '200':
          description: 查询成功
          schema:
            type: object
            properties:
              count:
                type: integer
              rows:
                type: array
                items:
                  $ref: '#/definitions/Staff'
    patch:
      summary: 更新多个员工
      description: 更新多个员工...
      parameters:
        - in: body
          name: body
          description: body参数
          schema:
            type: array
            items:
              $ref: '#/definitions/Staff'
      responses:
        '200':
          description: 更新成功
    delete:
      summary: 删除多个员工
      description: 删除多个员工...
      parameters:
        - in: body
          name: body
          description: body参数
          schema:
            type: object
            example:
              id:
                - '1'
                - '2'
      responses:
        '200':
          description: 更新成功
  '/staffs/{id}':
    get:
      summary: 查询单个员工
      description: 查询单个员工..
      parameters:
        - in: path
          name: id
          type: string
          required: true
          description: ID
      responses:
        '200':
          description: 查询成功
          schema:
            $ref: '#/definitions/Staff'
    patch:
      summary: 修改单个员工
      description: 修改单个员工..
      parameters:
        - in: path
          name: id
          type: string
          required: true
          description: ID
        - in: body
          name: body
          description: 模型参数
          required: true
          schema:
            $ref: '#/definitions/Staff'
      responses:
        '200':
          description: 修改成功
    delete:
      summary: 删除单个员工
      description: 删除单个员工..
      parameters:
        - in: path
          name: id
          type: string
          required: true
          description: ID
      responses:
        '200':
          description: 删除成功
  /staffs/multiple:
    post:
      summary: 创建多个员工
      description: 创建多个员工...
      parameters:
        - in: body
          name: body
          description: body参数
          schema:
            type: array
            items:
              $ref: '#/definitions/Staff'
      responses:
        '200':
          description: 创建成功
  /staff:
    get:
      summary: 查询单个员工
      description: 查询单个员工...
      parameters:
        - in: query
          name: id
          type: string
          description: 不限于id，可以是任意参数
          required: true
      responses:
        '200':
          description: 查询成功
          schema:
            $ref: '#/definitions/Staff'
    post:
      summary: 查找或创建单个员工
      description: 查找或创建单个员工...
      parameters:
        - in: body
          name: body
          description: 模型参数
          required: true
          schema:
            $ref: '#/definitions/Staff'
      responses:
        '200':
          description: 创建成功
          schema:
            $ref: '#/definitions/Staff'
...
definitions:
  Staff:
    type: object
    x-isModel: true
    x-plural: Staffs
    x-tableName: staff
    description: 员工
    required:
      - id
      - staff_code
      - name
    properties:
      id:
        type: string
        format: uuid
        description: ID
        uniqueItems: true
      user_id:
        type: string
        format: uuid
        description: 用户ID
        x-message: 请输入
        x-showTable: true
      staff_code:
        type: string
        format: string(64)
        description: 员工代码
        x-message: 请输入
        x-showTable: true
      name:
        type: string
        format: string(64)
        description: 企业名称
        x-message: 请输入
        x-showTable: true
      corp_code:
        type: string
        format: string(64)
        description: 企业代码
        x-message: 请输入
        x-showTable: true
      phone:
        type: string
        format: string(64)
        description: 电话
        x-message: 请输入
        x-showTable: true
      email:
        type: string
        format: email
        description: 邮箱
        x-message: 请输入
        x-showTable: true
      fox:
        type: string
        format: string(64)
        description: 传真
        x-message: 请输入
        x-showTable: true
    example:
      id: '123'
      staff_code: '1234'
      name: foo
...